resnetsoftmax/requirements.txt
streamlit
numpy
torch
torchvision
Pillow
scikit-learn
matplotlib
seaborn
tqdm